backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
pyjwt
cryptography
groq
pymysql
python-dotenv